% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotEnrich.R
\name{barplotEnrich}
\alias{barplotEnrich}
\title{Barplot the enrichResult}
\usage{
barplotEnrich(height, x = "Count", colorBy = "p.adjust", top = 5,
font.size = 12, title = "", color = NULL,...)
}
\arguments{
\item{height}{enrichResult object, alternatively output from \code{\link{functionEnrich}} function.}

\item{x}{one of "Count" and "GeneRatio" to specify the x axis.}

\item{colorBy}{one of 'pvalue', 'p.adjust', 'qvalue'.}

\item{top}{number of top categories to show.}

\item{font.size, title}{font size and title.}

\item{color}{the color of the bar.}

\item{...}{see parameters in \code{\link[ggplot2]{fortify}} function.}
}
\description{
This function is revised from the \code{\link[DOSE]{barplot.enrichResult}} function DOSE package, and used to
perform a barplot of the enrichResult object.
}

