% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{cate.gene}
\alias{cate.gene}
\title{Categorized genes}
\format{A list}
\usage{
data(cate.gene)
}
\value{
A list with five gene catogories.
}
\description{
A list with five gene categories.
}
\keyword{datasets}

