% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{cate.ratio}
\alias{cate.ratio}
\title{Expression Difference ratios of categorized genes}
\format{A list}
\usage{
data(cate.ratio)
}
\value{
A list with ED ratios for five gene catogories.
}
\description{
A list with five data frames of gene ED ratios
}
\keyword{datasets}

