% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorizeGene.R
\name{categorizeGene}
\alias{categorizeGene}
\title{Gene Categorization}
\usage{
categorizeGene(expr, diffGene, from.sample,to.sample, target.sample)
}
\arguments{
\item{expr}{a data frame with expression for all genes in \code{diffGene}, alternatively output from \code{\link{diffGene}}
function.}

\item{diffGene}{a list of differential genes in three comparisons, alternatively output by \code{\link{diffGene}} function.}

\item{from.sample, to.sample, target.sample}{character to specify the name of initiating sample, derived sample and primary
sample during a cellular engineering, must be consistent with sample names in the \code{expr} data frame.}
}
\value{
A list with components:
a list of the five gene categories
a list of the ED ratios for the five gene categories.
}
\description{
This function categorizes differential genes of each pair-wise comparison among e.g. initiating A, derived B and primary C
samples during a cellular engineering, into five categories named from \emph{Reversed},\emph{Inactive},\emph{Insufficient},
\emph{Successful} and \emph{Over} representing the gene reprogrammed states, and calculates the ratio of expression difference (ED)
 between B and A to the ED between C and A.
}
\details{
Gene (g) categorization is achieved by considering the pair-wise comparisons (Expression Difference, ED eq. 1) among
the three types of samples and the ratio of such differences (EDg ratio, eq. 2).
\eqn{EDg(B, A) = average expression of g in B minus average expression of g in A}    (1)
\eqn{EDg ratio = EDg(B, A)/EDg(C, A)}                                            (2)

\emph{Reversed}, EDg(B,A) and EDg(B,C) are significantly differential, while EDg(C,A) is not limited by differential or not,
 EDg ratio is smaller than 0;
\emph{Inactive}, EDg(C,A) and EDg(B,C) are significantly differential, while EDg(B,A) is not differential;
\emph{Insufficient}, EDg(B,A) ,EDg(C,A) and EDg(B,C) are all significantly differential, EDg ratio is between 0 and 1;
\emph{Successful}, EDg(B,A) and EDg(C,A) are significantly differential, while EDg(B,C) is not differential;
\emph{Over}, EDg(B,A) and EDg(B,C) are significantly differential, while EDg(C,A) is not limited by differential or not,
EDg ratio is greater than 1.
For the \emph{Inactive} and \emph{Successful} categories, genes with bottom and top 5 percentage ED ratios are
removed to avoid the ambiguous categorization with \emph{Reversed}, \emph{Insufficient} or \emph{Over} categories.
}

