% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{diffgene.genes}
\alias{diffgene.genes}
\title{Differential genes for three comparisons}
\format{A list}
\usage{
data(diffgene.genes)
}
\value{
A list with differential genes in three comparisons
}
\description{
A list of three vectors with significantly differential genes among three comparisons.
}
\keyword{datasets}

