% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffGene.R
\name{diffGene}
\alias{diffGene}
\title{Gene Expression Differential Analysis}
\usage{
diffGene(expr, array = TRUE, fpkm = FALSE, counts =FALSE, method =c("limma","DESeq2"),
from.sample, to.sample, target.sample, filter = FALSE, filter.perc = 0.4,
padjust ="fdr", signif = TRUE, pvalue = 0.05)
}
\arguments{
\item{expr}{a data frame with gene expression data.}

\item{array, fpkm, counts}{logical, specifying the type of input gene expression data.}

\item{method}{differential analysis method, alternatively to "limma" and "DESeq2", default to "limma".
"DESeq2" can be chosen only when \code{counts} is \code{TRUE}.}

\item{from.sample, to.sample, target.sample}{character to specify the name of initiating sample, derived sample and primary
sample during a cellular engineering.}

\item{filter}{logical to indicate whether the genes need to be filtered when match the parameter \code{filter.perc}
, only applied to fpkm and counts data.}

\item{filter.perc}{a 0 to 1 number to specify the gene filter criteria by the percentage of samples with non-zero expression.
Only used to fpkm and counts data when \code{filter} is \code{TRUE}, and filter the genes with non-zero expression in less than
filter.perc samples.}

\item{padjust}{indicate the method to do p.value correction, default to "fdr". See \code{\link[stats]{p.adjust}}.}

\item{signif}{logical to indicate whether only the significantly differential genes are output, default to FALSE.}

\item{pvalue}{a cutoff p.value for the significant genes, default to 0.05, only used when \code{signif} is TRUE.}
}
\value{
A list with components :
a list with differential analysis result for each pair-wise comparison;
a list with differential gene names for each pair-wise comparison;
a data frame with filtered/unfiltered gene expression.
}
\description{
Identify the differentially expressed genes for each pair-wise comparison of given three types of samples.
}
\details{
This function can be applied on both microarray and RNA-seq data for differential analysis when one of the "array",
"fpkm", or "counts" is specified. It does differential analysis to each pair-wise sample comparison among the from.sample,
to.sample and target.sample.
}
\examples{
data(SandlerFPKM)

# differential expression analysis:
diffgene = diffGene(expr = SandlerFPKM, array=FALSE,  fpkm=TRUE,  counts=FALSE,
                   from.sample="DMEC", to.sample="rEChMPP", target.sample="CB",
                   filter=TRUE, filter.perc =0.4, pvalue = 0.05 )

# differential analysis results
diffgene.result = diffgene[[1]]
# differential genes
diffgene.genes = diffgene[[2]]
# filtered expression data
expr.filter = diffgene[[3]]
}

