% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment Analysis by Hypergeometric Distribution Test}
\usage{
enrichment(cate.gene, annotated.gene, background.gene, padjust.method = "fdr" )
}
\arguments{
\item{cate.gene}{a list of the five gene categories, alternatively output by \code{\link{categorizeGene}}.}

\item{annotated.gene}{a list of annotated gene sets which the \code{cate.gene} are enriched in.}

\item{background.gene}{vector of background genes, e.g. all genes screened by microarray or RNA-sequencing.}

\item{padjust.method}{correction method for enrichment p-values, one of "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none", default to "fdr", see details in \code{\link[stats]{p.adjust}}.}
}
\value{
A list of enrichment results for the five gene categories.
}
\description{
Enrichment Analysis by Hypergeometric Distribution Test
}

