% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{expr.filter}
\alias{expr.filter}
\title{Filtered expression data}
\format{A data frame}
\usage{
data(expr.filter)
}
\value{
A data frame with filtered gene expression
}
\description{
A data frame with filtered RNASeq FPKM data.
}
\keyword{datasets}

