% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionEnrich.R
\name{functionEnrich}
\alias{functionEnrich}
\title{Funtional Enrichment Analysis}
\usage{
functionEnrich(cate.gene, organism = "human", convert = TRUE, from = "SYMBOL", ont = "BP",
pAdjustMethod = "bonferroni", GO = TRUE, KEGG = FALSE, raw = FALSE)
}
\arguments{
\item{cate.gene}{a list of the five gene categories, alternatively output by \code{\link{categorizeGene}}.}

\item{organism}{a character of organism "human" or "mouse" to indicate the species of background genes.}

\item{convert}{logical to determine whether the gene ID should be converted to "ENTREZID", default to \code{TRUE}.}

\item{from}{the gene id type of input data, see}

\item{ont}{One of "MF", "BP", and "CC" subontologies, see \code{\link[clusterProfiler]{enrichGO}}.}

\item{pAdjustMethod}{correction method for p-value, one of "holm", "hochberg", "hommel", "bonferroni", "BH",
"BY", "fdr", "none", default to "fdr", see details in \code{\link[stats]{p.adjust}}.}

\item{GO}{logical to determine whether the functional enrichment is performed on Gene Ontology, default to \code{TRUE}.}

\item{KEGG}{logical to determine whether the functional enrichment is performed on KEGG pathways, default to \code{FALSE}.}

\item{raw}{logical to determine if the raw "enrichResult" is output, default to \code{FALSE} to output a summary of the "enrichResult".}
}
\value{
Function enrichment analysis results.
}
\description{
This function performs Gene Ontology (GO) and KEGG pathways functional enrichment for the five gene categories by calling
\pkg{clusterProfiler} package.
}

