% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{goenrich}
\alias{goenrich}
\title{GO enrichment results}
\format{A list with 5 data frames}
\usage{
data(goenrich)
}
\description{
A list of 5 data frames with the Gene ontology enrichment results for the
five gene categories.
}
\keyword{datasets}

