% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnPlot.R
\name{grnPlot}
\alias{grnPlot}
\title{Gene Regulatory Network Plot}
\usage{
grnPlot(grn.data, cate.gene, filter = TRUE, nodes = NULL, centrality.score,
col = NULL, main= NULL, vertex.label =NULL, vertex.label.dist = 0, vertex.label.font = 1,
vertex.label.cex = 0.5, vertex.label.color="black", edge.arrow.size = 0.4,
edge.color = "grey", layout ="layout_with_lgl", legend.labels = NULL, ...)
}
\arguments{
\item{grn.data}{a data frame with two columns named "TF" and "TG" to specify the genes as transcription regulators (TF)
and target genes being regulated (TG).}

\item{cate.gene}{a list of the five gene categories as nodes in the network, alternatively output by
\code{\link{categorizeGene}}.}

\item{filter}{logical to specify if the network is reduced to less nodes by a filter, logical to \code{TRUE}
for a clear visualization.}

\item{nodes}{a character vector of genes to kept in the network, only applied when \code{filter} is \code{TRUE}.}

\item{centrality.score}{a vector or data frame of centrality scores for genes in the network, alternatively calculated
by \code{\link{networkAnalyze}} function.}

\item{col}{colors of gene vertex in each \code{cate.gene}.}

\item{main}{title of the plot.}

\item{vertex.label, vertex.label.dist, vertex.label.font, vertex.label.cex, vertex.label.color}{parameters for
vertex labels, to specify the labels of vertex, the position of labels on the vertex, font size, cex and color
for label, see details in \code{\link[igraph]{igraph.plotting}}.}

\item{edge.arrow.size, edge.color}{parameters for edge to specify the arrow size and color of edge, see details
in \code{\link[igraph]{igraph.plotting}}.}

\item{layout}{the layout of network plot, see details in \code{\link[igraph]{layout}}.}

\item{legend.labels}{vector of label names for each \code{cate.gene}.}

\item{...}{other parameters used in \code{\link[igraph]{igraph.plotting}}.}
}
\value{
a igraph plot for gene regulatory network.
}
\description{
This function plot the a cell/tissue-specific gene regulatory network with genes in the five categories,
alterntively reducing the network size by plotting given specific genes as nodes.
}
\examples{
\dontrun{
  # select genes to shown their regulation with others
  node.genes = c("ZNF641", "BCL6")
  # enlarge the centrality
  centrality.score = degree$centrality*100
  names(centrality.score) = degree$Gene
  par(mar = c(2,2,3,2))
  grnPlot(grn.data = human.grn[[tissue]], cate.gene = cate.gene, filter = TRUE,
         nodes = node.genes, centrality.score = centrality.score,
         main = "Gene regulatory network")
}
}

