% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlot.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Heatmap Plot of Enriched Terms}
\usage{
heatmapPlot(enrichresult, GO = FALSE, terms = NULL, padjust = TRUE, pvalue = 0.05,
top= NA, filter = FALSE, main = NA, annotation = NULL, annotation_col = NULL,
annotated_row = FALSE, annotation_row = NULL,annotation_colors = NA,
display_numbers = FALSE, annotation_legend = FALSE,...)
}
\arguments{
\item{enrichresult}{a list of data frames with enrichment results, alternatively output by \code{\link{functionEnrich}}
or \code{\link{enrichment}}.}

\item{GO}{logical to determine whether the terms are Gene Ontology(GO) terms enriched by \code{\link{functionEnrich}}.}

\item{terms}{a character vector to specify the terms chosen to be listed in the heatmap, selected from the \code{enrichment} result,
or a data frame with terms and corresponding term annotations used for \code{annotation_row}.}

\item{padjust}{logical to determine whether the significantly enriched terms were selected by adjusted p.value
rather than the p.value, default to \code{TRUE}.}

\item{pvalue}{a cutoff value to select the significantly enriched terms.}

\item{top}{a number to specify the most significantly enriched terms to be drawn in each category, default to \code{NA}
without specifying.}

\item{filter}{logical to specify whether the enriched terms need to be filered with the ones which are significant among
the first four categories.}

\item{main}{a character of main title on the heatmap plot.}

\item{annotation, annotation_row, annotation_col, annotation_colors, annotation_legend, ...}{see details in \code{\link[pheatmap]{pheatmap}}.}

\item{annotated_row}{logical to determine whether the the rows are annotated by \code{annotation_row}, default to \code{FALSE}.
When it's \code{TRUE}, \code{annotation_row} should be specified or using annotations in a data frame of \code{terms}.}

\item{display_numbers}{logical to determine whether the gene counts number is shown on the heatmap.}
}
\value{
heatmap plot and the terms with p.values for the heatmap
}
\description{
This function plot the significantly enriched terms in a heatmap by calling \pkg{pheatmap} package.
}
\examples{
# plot the enrichment results by the five gene categories
data(goenrich)
heatmaptable = heatmapPlot(goenrich, GO = TRUE, top = 5, filter = FALSE,
                           main = "Gene ontology enrichment",
                           display_numbers =  FALSE)
}

