% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{human.gene}
\alias{human.gene}
\title{Gene regulatory network based human cell/tissue-specific gene sets}
\format{A list}
\usage{
data(human.gene)
}
\description{
A list of 16 human cells/tissues-specific gene sets summarized from the gene regulatory network
downloaded from the CellNet website.
}
\keyword{datasets}

