% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{human.grn}
\alias{human.grn}
\title{Human cell/tissue-specific gene-gene regulation}
\format{A list with 16 data frames}
\usage{
data(human.grn)
}
\value{
A list of human gene regulatory information.
}
\description{
A list of 16 data frames (cells/tissues) with transcription factors (TF) to target genes (TG)
regulation information.
}
\keyword{datasets}

