% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{human.tf}
\alias{human.tf}
\title{Gene regulatory network based human cell/tissue-specific transcription factor (TF)
regulated gene sets}
\format{A list}
\usage{
data(human.tf)
}
\description{
A list of 1455 human cells/tissues-specific TF regulated gene sets summarized from the gene
regulatory network downloaded from the CellNet website.
}
\keyword{datasets}

