% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{markers}
\alias{markers}
\title{Marker genes}
\format{A vector}
\usage{
data(markers)
}
\description{
A vector containing 65 genes specific in endothelial and haematopoietic cells as
 listed in Sandler's paper.
}
\keyword{datasets}

