% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerScatter.R
\name{markerScatter}
\alias{markerScatter}
\title{Scatter Plot for Gene Expression}
\usage{
markerScatter(expr, log = FALSE, samples, cate.gene, markers, pch = 19, cex = 0.5,
col = NULL, xlab = NULL, ylab = NULL, main = NULL, add.line = TRUE,text.cex = 1, legend.labels = NULL,
...  )
}
\arguments{
\item{expr}{a data frame with gene expression.}

\item{log}{logical to determine if the gene expression data is log converted (add a small constant 2), default to FALSE.}

\item{samples}{a vector of samples to compare on the x axis and y axis.}

\item{cate.gene}{a list of the gene categories, alternatively output by \code{\link{categorizeGene}}.}

\item{markers}{vector of marker genes to be highlighted in the plot. No gene is highlighted when it's \code{NULL}.}

\item{pch, cex, col, xlab, ylab, main}{plot parameters, see details in \code{\link[graphics]{par}}.}

\item{add.line}{logical to determine if the linear model fitting line is added on the figure.}

\item{text.cex}{font size for the text on \code{markers}, see details in \code{\link[graphics]{text}}.}

\item{legend.labels}{vector of labels for the marker legend.}

\item{...}{other parameters in \code{\link[graphics]{plot}}.}
}
\value{
plot with gene expression profile.
}
\description{
Generates an expression profile of each gene catetory in one sample against another, alternatively plot the regression
line from linear modeling fitting.
}
\details{
Visualization of gene expression in the five categories under each pair-wised comparison.
}
\examples{
#load the marker genes of somatic and primary cells
data(markers)
data(expr.filter)
#scatterplot
col = c("#abd9e9", "#2c7bb6", "#fee090", "#d7191c", "#fdae61")
markerScatter(expr = expr.filter, log = TRUE, samples = c("CB", "DMEC"),
              cate.gene = cate.gene[2:4], markers = markers, col = col[2:4],
              xlab = expression('log'[2]*' expression in CB (target)'),
             ylab = expression('log'[2]*' expression in DMEC (input)'),main = "")
}

