% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{mouse.gene}
\alias{mouse.gene}
\title{Gene regulatory network based mouse cell/tissue-specific gene sets}
\format{A list}
\usage{
data(mouse.gene)
}
\value{
A list of 20 mouse cells/tissues-specific gene sets
}
\description{
A list of 20 mouse cells/tissues-specific gene sets summarized from the gene regulatory network
downloaded from the CellNet website.
}
\keyword{datasets}

