% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{mouse.grn}
\alias{mouse.grn}
\title{Mouse cell/tissue-specific gene-gene regulation}
\format{A list with 20 data frames}
\usage{
data(mouse.grn)
}
\description{
A list of 20 data frames (cells/tissues) with transcription factors (TF) to target genes (TG)
regulation information.
}
\keyword{datasets}

