% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{mouse.tf}
\alias{mouse.tf}
\title{Gene regulatory network based mouse cell/tissue-specific transcription factor (TF) regulated
gene sets}
\format{A list}
\usage{
data(mouse.tf)
}
\value{
A list of 1744 mouse cells/tissues-specific TF regulated gene sets.
}
\description{
A list of 1744 mouse cells/tissues-specific TF regulated gene sets summarized from the gene
regulatory network downloaded from the CellNet website.
}
\keyword{datasets}

