% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAnalyze.R
\name{networkAnalyze}
\alias{networkAnalyze}
\title{Network Topological Analysis}
\usage{
networkAnalyze(grn.data, cate.gene, centrality = c("degree", "betweenness",
"stress", "closeness"), mode = c("all","in", "out", "total"))
}
\arguments{
\item{grn.data}{a data frame with two columns named "TF" and "TG" to specify the genes as transcription regulators (TF)
and target genes (TG) being regulated.}

\item{cate.gene}{a list of the five gene categories as nodes in the network, alternatively output by
\code{\link{categorizeGene}}.}

\item{centrality}{charactor string of "degree", "betweenness", "closeness" and "stress" to calculate the
centrality of network built from input \code{grn.data}, see \code{\link[igraph]{degree}}.
\code{\link[igraph]{betweenness}}, \code{\link[igraph]{closeness}}, and \code{\link[sna]{stresscent}} for details.}

\item{mode}{character string of "all", "in", "out" and "total", only used when \code{centrality} is "degree"
or "closeness", see \code{\link[igraph]{degree}}, \code{\link[igraph]{closeness}} for details.}
}
\value{
data frame with genes and centrality scores.
}
\description{
This function analyzes the topological structure of gene regulation network (GRN) by calculating the "degree",
"betweenness", "closeness" and "stress" parameters, and output the centrality values for given genes in each
gene categories.
}
\examples{
# load the CellNet GRN and gene categories
data(human.grn)
data(cate.gene)

# specify a tissue-specifc network
tissue = "Hspc"
degree = networkAnalyze(human.grn[[tissue]], cate.gene = cate.gene,
                       centrality = "degree", mode ="all")
}

