% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{SandlerFPKM}
\alias{SandlerFPKM}
\title{RNA-seq data in FPKM
A data frame containing 16692 gene in 7 samples from paper published by Sandler et.al. in 2014.,
samples are named from somatic dermal microvascular endothelial cell (DMEC) to induced
multipotent haematopoietic progenitor (rEChMPP) cells and primary cord blood (CB) cells.}
\format{A data frame with 16692 rows and 7 columns specifying for genes and samples.}
\usage{
data(SandlerFPKM)
}
\value{
A data frame with gene expression data.
}
\description{
RNA-seq data in FPKM
A data frame containing 16692 gene in 7 samples from paper published by Sandler et.al. in 2014.,
samples are named from somatic dermal microvascular endothelial cell (DMEC) to induced
multipotent haematopoietic progenitor (rEChMPP) cells and primary cord blood (CB) cells.
}
\keyword{datasets}

