% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{tissueGenes}
\alias{tissueGenes}
\title{Cell/Tissue specific gene sets}
\format{A list}
\usage{
data(tissueGenes)
}
\value{
A list with 126 cells/tissues-specific gene sets
}
\description{
A list of 126 cells/tissues-specific gene sets identified by SpeCond algorithm
 from 126 cells/tissues in Gene Enrichment Profiler database.
}
\keyword{datasets}

