% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegc.R
\docType{data}
\name{tissueGroup}
\alias{tissueGroup}
\title{Tissue mapping to groups}
\format{A data frame with 126 rows (tissues) and 3 columns (Tissue, Tissue_abbr, Group)}
\usage{
data(tissueGroup)
}
\description{
A data frame with 126 cells/tissues and 30 C/T groups they belong to.
}
\keyword{datasets}

