% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerScatter.R
\name{markerScatter}
\alias{markerScatter}
\title{Scatter Plot for Gene Expression}
\usage{
markerScatter(expr, log = FALSE, samples, cate.gene, markers, pch = 19, cex = 0.5,
col = NULL, xlab = NULL, ylab = NULL, main = NULL, add.line = TRUE, legend.labels = NULL,
...  )
}
\arguments{
\item{expr}{a data frame with gene expression.}

\item{log}{logical to determine if the gene expression data is log converted (add a small constant 2), default to FALSE.}

\item{samples}{a vector of samples to compare on the x axis and y axis.}

\item{cate.gene}{a list of the gene categories, alternatively output by \code{\link{categorizeGene}}.}

\item{markers}{vector of marker genes to be highlighted in the plot. No gene is highlighted when it's \code{NULL}.}

\item{pch, cex, col, xlab, ylab, main}{plot parameters, see details in \code{\link[graphics]{par}}.}

\item{add.line}{logical to determine if the linear model fitting line is added on the figure.}

\item{legend.labels}{vector of labels for the marker legend.}

\item{...}{other parameters in \code{\link[graphics]{plot}}.}
}
\description{
Generates an expression profile of each gene catetory in one sample against another, alternatively plot the regression
line from linear modeling fitting.
}
\details{
Visualization of gene expression in the five categories under each pair-wised comparison.
}

